+++ pkg/distributor/ha_tracker.go
@@ -143,8 +143,6 @@ type HATrackerConfig struct {
 	EnableElectedReplicaMetric bool `yaml:"enable_elected_replica_metric"`
 
 	KVStore kv.Config `yaml:"kvstore" doc:"description=Backend storage to use for the ring. Supported values are: consul, etcd, inmemory, memberlist, multi. Note that etcd is deprecated."`
-
-	DeprecatedHATrackerTimeoutsConfig `yaml:",inline"`
 }
 
 // RegisterFlags adds the flags required to config this to the given FlagSet.
@@ -157,16 +155,6 @@ func (cfg *HATrackerConfig) RegisterFlags(f *flag.FlagSet) {
 	// order to not clash with the ring key if they both share the same KVStore
 	// backend (i.e. run on the same Consul cluster).
 	cfg.KVStore.RegisterFlagsWithPrefix("distributor.ha-tracker.", "ha-tracker/", f)
-}
-
-// DeprecatedHATrackerTimeoutsConfig is kept for backwards-compatibility in the
-// YAML config files. Values are copied to limits config, and must be accesed
-// through haTrackerLimits.HATrackerTimeouts.
-// TODO: Remove in Mimir 2.18.0
-type DeprecatedHATrackerTimeoutsConfig struct {
-	DeprecatedUpdateTimeout          time.Duration `yaml:"ha_tracker_update_timeout" category:"advanced" doc:"nocli|description=Deprecated. Use limits.ha_tracker_update_timeout."`
-	DeprecatedUpdateTimeoutJitterMax time.Duration `yaml:"ha_tracker_update_timeout_jitter_max" category:"advanced" doc:"nocli|description=Deprecated. Use limits.ha_tracker_update_timeout_jitter_max."`
-	DeprecatedFailoverTimeout        time.Duration `yaml:"ha_tracker_failover_timeout" category:"advanced" doc:"nocli|description=Deprecated. Use limits.ha_tracker_failover_timeout."`
 }
 
 func GetReplicaDescCodec() codec.Proto {