+++ pkg/distributor/push_test.go
@@ -42,6 +42,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/prometheus/common/model"
+	"github.com/prometheus/otlptranslator"
 	"github.com/prometheus/prometheus/prompb"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -562,10 +563,11 @@ func createMalformedRW1Request(t testing.TB, protobuf []byte) *http.Request {
 	return req
 }
 
-func createPrometheusRemoteWriteProtobuf(t testing.TB) []byte {
+func createPrometheusRemoteWriteProtobuf(t testing.TB, timeseries ...prompb.TimeSeries) []byte {
 	t.Helper()
-	input := prompb.WriteRequest{
-		Timeseries: []prompb.TimeSeries{
+
+	if len(timeseries) == 0 {
+		timeseries = []prompb.TimeSeries{
 			{
 				Labels: []prompb.Label{
 					{Name: "__name__", Value: "foo"},
@@ -576,29 +578,36 @@ func createPrometheusRemoteWriteProtobuf(t testing.TB) []byte {
 				Histograms: []prompb.Histogram{
 					prompb.FromIntHistogram(1337, test.GenerateTestHistogram(1))},
 			},
-		},
+		}
+	}
+
+	input := prompb.WriteRequest{
+		Timeseries: timeseries,
 	}
 	inputBytes, err := input.Marshal()
 	require.NoError(t, err)
 	return inputBytes
 }
 
-func createMimirWriteRequestProtobuf(t *testing.T, skipLabelNameValidation, skipLabelCountValidation bool) []byte {
+func createMimirWriteRequestProtobuf(t *testing.T, skipLabelNameValidation, skipLabelCountValidation bool, timeseries ...mimirpb.PreallocTimeseries) []byte {
 	t.Helper()
 	h := prompb.FromIntHistogram(1337, test.GenerateTestHistogram(1))
-	ts := mimirpb.PreallocTimeseries{
-		TimeSeries: &mimirpb.TimeSeries{
-			Labels: []mimirpb.LabelAdapter{
-				{Name: "__name__", Value: "foo"},
+
+	if len(timeseries) == 0 {
+		timeseries = []mimirpb.PreallocTimeseries{{
+			TimeSeries: &mimirpb.TimeSeries{
+				Labels: []mimirpb.LabelAdapter{
+					{Name: "__name__", Value: "foo"},
+				},
+				Samples: []mimirpb.Sample{
+					{Value: 1, TimestampMs: time.Date(2020, 4, 1, 0, 0, 0, 0, time.UTC).UnixNano()},
+				},
+				Histograms: []mimirpb.Histogram{promToMimirHistogram(&h)},
 			},
-			Samples: []mimirpb.Sample{
-				{Value: 1, TimestampMs: time.Date(2020, 4, 1, 0, 0, 0, 0, time.UTC).UnixNano()},
-			},
-			Histograms: []mimirpb.Histogram{promToMimirHistogram(&h)},
-		},
+		}}
 	}
 	input := mimirpb.WriteRequest{
-		Timeseries:               []mimirpb.PreallocTimeseries{ts},
+		Timeseries:               timeseries,
 		Source:                   mimirpb.RULE,
 		SkipLabelValidation:      skipLabelNameValidation,
 		SkipLabelCountValidation: skipLabelCountValidation,
@@ -1427,7 +1436,6 @@ func TestHandler_EnforceInflightBytesLimitInfluxPush(t *testing.T) {
 }
 
 func TestHandler_EnforceInflightBytesLimitOTLP(t *testing.T) {
-
 	jsonReq := &httpgrpc.HTTPRequest{
 		Method: "POST",
 		Headers: []*httpgrpc.Header{
@@ -1856,7 +1864,13 @@ func (o otlpLimitsMock) OTelPromoteScopeMetadata(string) bool {
 
 func (o otlpLimitsMock) OTelNativeDeltaIngestion(string) bool { return false }
 
-func (o otlpLimitsMock) OTelEnableUnescapedNames(string) bool { return false }
+func (o otlpLimitsMock) OTelTranslationStrategy(string) otlptranslator.TranslationStrategyOption {
+	return otlptranslator.UnderscoreEscapingWithoutSuffixes
+}
+
+func (o otlpLimitsMock) NameValidationScheme(string) model.ValidationScheme {
+	return model.LegacyValidation
+}
 
 func promToMimirHistogram(h *prompb.Histogram) mimirpb.Histogram {
 	pSpans := make([]mimirpb.BucketSpan, 0, len(h.PositiveSpans))